harga to integer

+++ database/migrations/2021_12_04_155008_create_wisatas_table.php
@@ -19,7 +19,7 @@ class CreateWisatasTable extends Migration
             $table->string('lokasi');
             $table->string('deskripsi');
             $table->string('url_gambar');
-            $table->double('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
